+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Changes since last release]
+### Added
+- Support for clustered box plots.
 ### Fixed
 - Bug in the spacing and sizing of bars in clustered bar charts.